+++ src/DeviceOperator.ts
@@ -5,17 +5,11 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import * as fs from 'fs-plus';
-import * as path from 'path';
 import {IoTProject} from './Models/IoTProject';
 import {TelemetryContext} from './telemetry';
 import {Board, BoardQuickPickItem} from './Models/Interfaces/Board';
 import {ArduinoPackageManager} from './ArduinoPackageManager';
 import {BoardProvider} from './boardProvider';
-import {ConfigHandler} from './configHandler';
-import {ConfigKey} from './constants';
-import {OTA} from './Models/OTA';
-import {askAndNewProject} from './utils';
 
 export class DeviceOperator {
   async compile(
@@ -97,91 +91,6 @@ export class DeviceOperator {
 
     vscode.window.showInformationMessage(
         `Device package for ${boardSelection.label} has been installed.`);
-    return true;
-  }
-
-  async generateCrc(
-      context: vscode.ExtensionContext, channel: vscode.OutputChannel,
-      telemetryContext: TelemetryContext) {
-    if (!vscode.workspace.workspaceFolders) {
-      vscode.window.showWarningMessage('No workspace opened.');
-      channel.show();
-      channel.appendLine('No workspace opened.');
-      return false;
-    }
-
-    const devicePath = ConfigHandler.get<string>(ConfigKey.devicePath);
-    if (!devicePath) {
-      vscode.window.showWarningMessage(
-          'No device path found in workspace configuration.');
-      channel.show();
-      channel.appendLine('No device path found in workspace configuration.');
-      return false;
-    }
-    const deviceBuildLocation = path.join(
-        vscode.workspace.workspaceFolders[0].uri.fsPath, '..', devicePath,
-        '.build');
-
-    if (!deviceBuildLocation) {
-      vscode.window.showWarningMessage(
-          'No device compile output folder found.');
-      channel.show();
-      channel.appendLine('No device compile output folder found.');
-      return false;
-    }
-
-    const binFiles = fs.listSync(deviceBuildLocation, ['bin']);
-    if (!binFiles || !binFiles.length) {
-      const message =
-          'No bin file found. Please run the command of Device Compile first.';
-      vscode.window.showWarningMessage(message);
-      channel.show();
-      channel.appendLine(message);
-      return false;
-    }
-
-    let binFilePath = '';
-
-    if (binFiles.length === 1) {
-      binFilePath = binFiles[0];
-    } else {
-      const binFilePickItems: vscode.QuickPickItem[] = [];
-      for (const file of binFiles) {
-        const fileName = path.basename(file);
-        binFilePickItems.push({label: fileName, description: file});
-      }
-
-      const choice = await vscode.window.showQuickPick(binFilePickItems, {
-        ignoreFocusOut: true,
-        matchOnDescription: true,
-        matchOnDetail: true,
-        placeHolder: 'Select bin file',
-      });
-
-      if (!choice || !choice.description) {
-        return false;
-      }
-
-      binFilePath = choice.description;
-    }
-
-    if (!binFilePath || !fs.existsSync(binFilePath)) {
-      return false;
-    }
-
-    const res = OTA.generateCrc(binFilePath);
-
-    vscode.window.showInformationMessage('Generate CRC succeeded.');
-
-    channel.show();
-    channel.appendLine('========== CRC Information ==========');
-    channel.appendLine('');
-    channel.appendLine('fwPath: ' + binFilePath);
-    channel.appendLine('fwPackageCheckValue: ' + res.crc);
-    channel.appendLine('fwSize: ' + res.size);
-    channel.appendLine('');
-    channel.appendLine('======================================');
-
     return true;
   }
 }